+++ src/Conflux.js
@@ -658,6 +658,12 @@ class Conflux {
    * - outcomeStatus `number`:  the outcome status code, 0 was successful, 1 for an error occurred in the execution.
    * - logsBloom `string`: Bloom filter for light clients to quickly retrieve related logs.
    * - logs `object[]`: Array of log objects, which this transaction generated.
+   * - gasCoveredBySponsor `boolean`: `true` if this transaction's gas fee was covered by the sponsor.
+   * - storageCoveredBySponsor `boolean`: `true` if this transaction's storage collateral was covered by the sponsor.
+   * - storageCollateralized `BigInt`: the amount of storage collateral this transaction required.
+   * - storageReleased `array`: array of storage change objects, each specifying an address and the corresponding amount of storage collateral released
+   *   - address `string`: address released
+   *   - collaterals `BigInt`: corresponding amount of storage collateral released
    *
    * @example
    * > await conflux.getTransactionReceipt('0xbf7110474779ba2404433ef39a24cb5b277186ef1e6cb199b0b60907b029a1ce');
@@ -675,7 +681,14 @@ class Conflux {
       stateRoot: '0xd6a7c2c14cb0d1233010acca98e114db5a10e0b94803d23b01a6777b7fd3b2fd',
       to: '0x83bf953c8b687f0d1b8d2243a3e0654ec1f70d1b',
       transactionHash: '0xbf7110474779ba2404433ef39a24cb5b277186ef1e6cb199b0b60907b029a1ce',
-      txExecErrorMsg: null
+      txExecErrorMsg: null,
+      gasCoveredBySponsor: false,
+      storageCoveredBySponsor: false,
+      storageCollateralized: 0n,
+      storageReleased: [
+        address: '0x0000000000000000000000000000000000000001',
+        collaterals: 640n,
+      ],
     }
    */
   async getTransactionReceipt(transactionHash) {